Fix incorrect behaviors inside the Google Ads stats card
